soc: ti: k3: add mpu configuration for am2434
Adds a default MPU configuration based on the TI MCU+ SDK.

The SRAM region the core runs on needs to be marked as non-shareable since
atomic instructions are used on it and for this SoC it's connected in a way
that dosen't allow atomic instructions (e.g. ldrex) to have exclusive
access.

Due to that the Cortex-R causes a data abort in such cases. To allow these
instructions anyway the region is marked a non-shared which is fine as long
as only one core is accessing the memory area without external
synchronization.

Signed-off-by: Mika Braunschweig <mika.braunschweig@siemens.com>