drivers: mspi: ti: k3: add initial MSPI driver
Add a pretty basic MSPI driver. The driver was tested in 1S-1S-1S mode on a
am243x-lp board with the onboard Infineon flash. Theoretically the driver
also has support for other MSPI modes though they weren't tested. The
command and dummy cycles are always taken from the xfer request and never
from the devicetree.

The driver is pretty basic for now and lacks e.g. callback implementation.
This is something that can be added in the future. If a non-supported /
invalid request is detected a error code is returned.

Signed-off-by: Mika Braunschweig <mika.braunschweig@siemens.com>