drivers: mtd: add infineon s25h mspi nor flash
Add a infineon s25h mspi nor flash driver. This driver was tested on the
am243x-lp board in 1S-1S-1S mode.

The driver doesn't check the configuration for dummy cycles and assumes a
factory default configuration. It also changes, if needed, a non-volatile
configuration bit so the erase sectors are uniform. On startup the JEDEC id
is read and checked as basic communication test.

Signed-off-by: Mika Braunschweig <mika.braunschweig@siemens.com>