soc: ti: k3: split up soc-specific Kconfig
Split up Kconfig settings for the ti Keystone 3 (K3) architecture. The old
configuration assumed that every R5F is initialized via a table that's
compatible with the linux remote proc framework and OpenAMP library.

Additionally the number of interrupts and speed of the system timer was
specified only by the cpu type for this series.

For better extendability this was split up into different files that
reflect the subseries of the soc.

Additionally this sets the number of interrupts and the clock frequency for
the R5F in the AM243x to the correct value and the MPU is enabled by
default.

Signed-off-by: Mika Braunschweig <mika.braunschweig@siemens.com>