arch: arm: cortex_a_r: add option to boot other CPUs into loop
Some SoCs start multiple CPU cores in a SMP configuration but sometimes
there is missing support in Zephyr or the SoCs has requirements that can't
be fulfilled by Zephyr yet (e.g. all spinlocks need to be implemented using
hardware spinlocks since the sram doesn't support exclusive access).

To address this an option was added that puts all CPU cores except the one
with the number 0 (according to the Multiprocessor Affinity Register with
the bits for Affinity Level 0) into an endless loop directly on startup.

Signed-off-by: Mika Braunschweig <mika.braunschweig@siemens.com>