soc: ti: k3: add support for loading zephyr directly
This commit adds all required parts to boot a Zephyr application directly
on the AM2434 Launchpad. Additionally it has been verified that MCUboot can
be used to boot a Zephyr application from the flash.

To achieve this a new variant has been added that specifies the application
is built to run directly after the ROM bootloader. This will automatically
generate a binary that can be loaded by the ROM bootloader that's running
on the chip.

Signed-off-by: Mika Braunschweig <mika.braunschweig@siemens.com>