UPSTREAM-PEND: drivers: mspi: cdns: add initial MSPI driver
Add a Cadence MSPI peripheral driver, used in the TI K3 platform. The
driver was tested in 1S-1S-1S and 4S-4S-4S mode with the onboard infineon
s25h flash of the am243x launchpad and a custom driver for the flash using
this interface.
The command and dummy cycles are always taken from the xfer request and
never from the devicetree since different commands might have different
latencies.
The driver is somewhat basic for now and lacks e.g. callback
implementation. This is something that can be added in the future. If a
non-supported / invalid request is detected a error code is returned.

PR: 102513

Signed-off-by: Mika Braunschweig <mika.braunschweig@siemens.com>